Correct int types in Netjack2
git-svn-id: http://subversion.jackaudio.org/jack/jack2/trunk/jackmp@2505 0c269be4-1314-0410-8aa9-9f06e86f4224

+++ common/JackNetDriver.h
@@ -32,10 +32,10 @@ namespace Jack
 		private:
 			session_params_t fParams;
 			char* fMulticastIP;
-			unsigned int fPort;
+			size_t fPort;
 			int fSockfd;
 			struct sockaddr_in fMasterAddr;
-			unsigned int fNSubProcess;
+			size_t fNSubProcess;
 
 			jack_port_id_t* fMidiCapturePortList;
 			jack_port_id_t* fMidiPlaybackPortList;
@@ -67,23 +67,23 @@ namespace Jack
 			JackMidiBuffer* GetMidiInputBuffer ( int port_index );
 			JackMidiBuffer* GetMidiOutputBuffer ( int port_index );
 
-			int Recv ( unsigned int size, int flags );
-			int Send ( unsigned int size, int flags );
-			
+			int Recv ( size_t size, int flags );
+			int Send ( size_t size, int flags );
+
 		public:
 			JackNetDriver ( const char* name, const char* alias, JackLockedEngine* engine, JackSynchro* table,
-				const char* ip, unsigned int port, int midi_input_ports, int midi_output_ports, const char* master_name );
+			                const char* ip, size_t port, int midi_input_ports, int midi_output_ports, const char* master_name );
 			~JackNetDriver();
 
 			int Open ( jack_nframes_t frames_per_cycle, jack_nframes_t rate, bool capturing, bool playing,
 			           int inchannels, int outchannels, bool monitor, const char* capture_driver_name,
 			           const char* playback_driver_name, jack_nframes_t capture_latency, jack_nframes_t playback_latency );
-            
-            int Attach();
-            int Detach();
-            
-            int Read();
-            int Write();
+
+			int Attach();
+			int Detach();
+
+			int Read();
+			int Write();
 	};
 }
