+++ common/JackNetTool.cpp
@@ -8,7 +8,7 @@ namespace Jack
 {
 // NetMidiBuffer**********************************************************************************
 
-	NetMidiBuffer::NetMidiBuffer ( session_params_t* params, unsigned int nports, char* net_buffer )
+	NetMidiBuffer::NetMidiBuffer ( session_params_t* params, size_t nports, char* net_buffer )
 	{
 		fNPorts = nports;
 		fMaxBufsize = fNPorts * sizeof ( sample_t ) * params->fPeriodSize ;
@@ -26,7 +26,7 @@ namespace Jack
 		delete[] fPortBuffer;
 	}
 
-	unsigned int NetMidiBuffer::GetSize()
+	size_t NetMidiBuffer::GetSize()
 	{
 		return fMaxBufsize;
 	}
@@ -35,7 +35,7 @@ namespace Jack
 	{
 		for ( int port_index = 0; port_index < fNPorts; port_index++ )
 		{
-			for ( unsigned int event = 0; event < fPortBuffer[port_index]->event_count; event++ )
+			for ( size_t event = 0; event < fPortBuffer[port_index]->event_count; event++ )
 				if ( fPortBuffer[port_index]->IsValid() )
 					jack_info ( "port %d : midi event %u/%u -> time : %u, size : %u",
 						port_index + 1, event + 1, fPortBuffer[port_index]->event_count,
@@ -75,13 +75,13 @@ namespace Jack
 		return pos;
 	}
 
-	int NetMidiBuffer::RenderFromNetwork ( unsigned int subcycle, unsigned int copy_size )
+	int NetMidiBuffer::RenderFromNetwork ( size_t subcycle, size_t copy_size )
 	{
 		memcpy ( fBuffer + subcycle * fMaxPcktSize, fNetBuffer, copy_size );
 		return copy_size;
 	}
 
-	int NetMidiBuffer::RenderToNetwork ( unsigned int subcycle, unsigned int total_size )
+	int NetMidiBuffer::RenderToNetwork ( size_t subcycle, size_t total_size )
 	{
 		int size = total_size - subcycle * fMaxPcktSize;
 		int copy_size = ( size <= fMaxPcktSize ) ? size : fMaxPcktSize;
@@ -91,7 +91,7 @@ namespace Jack
 
 // net audio buffer *********************************************************************************
 
-	NetAudioBuffer::NetAudioBuffer ( session_params_t* params, unsigned int nports, char* net_buffer )
+	NetAudioBuffer::NetAudioBuffer ( session_params_t* params, size_t nports, char* net_buffer )
 	{
 		fNPorts = nports;
 		fPeriodSize = params->fPeriodSize;
@@ -108,18 +108,18 @@ namespace Jack
 		delete[] fPortBuffer;
 	}
 
-	unsigned int NetAudioBuffer::GetSize()
+	size_t NetAudioBuffer::GetSize()
 	{
 		return fNPorts * fSubPeriodBytesSize;
 	}
 
-	void NetAudioBuffer::RenderFromJackPorts ( unsigned int subcycle )
+	void NetAudioBuffer::RenderFromJackPorts ( size_t subcycle )
 	{
 		for ( int port_index = 0; port_index < fNPorts; port_index++ )
 			memcpy ( fNetBuffer + port_index * fSubPeriodBytesSize, fPortBuffer[port_index] + subcycle * fSubPeriodSize, fSubPeriodBytesSize );
 	}
 
-	void NetAudioBuffer::RenderToJackPorts ( unsigned int subcycle )
+	void NetAudioBuffer::RenderToJackPorts ( size_t subcycle )
 	{
 		for ( int port_index = 0; port_index < fNPorts; port_index++ )
 			memcpy ( fPortBuffer[port_index] + subcycle * fSubPeriodSize, fNetBuffer + port_index * fSubPeriodBytesSize, fSubPeriodBytesSize );
@@ -257,24 +257,24 @@ namespace Jack
 
 // Utility *******************************************************************************************************
 
-	EXPORT unsigned int SetFramesPerPacket ( session_params_t* params )
+	EXPORT size_t SetFramesPerPacket ( session_params_t* params )
 	{
 		if ( !params->fSendAudioChannels && !params->fReturnAudioChannels )
 			return ( params->fFramesPerPacket = params->fPeriodSize );
-		unsigned int period = ( int ) powf ( 2.f, ( int ) log2 ( ( params->fMtu - sizeof ( packet_header_t ) )
+		size_t period = ( int ) powf ( 2.f, ( int ) log2 ( ( params->fMtu - sizeof ( packet_header_t ) )
 			/ ( max ( params->fReturnAudioChannels, params->fSendAudioChannels ) * sizeof ( sample_t ) ) ) );
 		( period > params->fPeriodSize ) ? params->fFramesPerPacket = params->fPeriodSize : params->fFramesPerPacket = period;
 		return params->fFramesPerPacket;
 	}
 
-	EXPORT unsigned int GetNMidiPckt ( session_params_t* params, unsigned int data_size )
+	EXPORT size_t GetNMidiPckt ( session_params_t* params, size_t data_size )
 	{
 		//even if there is no midi data, jack need an empty buffer to know there is no event to read
 		//99% of the cases : all data in one packet
 		if ( data_size <= ( params->fMtu - sizeof ( packet_header_t ) ) )
 			return 1;
 		//else, get the number of needed packets (simply slice the biiig buffer)
-		unsigned int npckt = data_size / ( params->fMtu - sizeof ( packet_header_t ) );
+		size_t npckt = data_size / ( params->fMtu - sizeof ( packet_header_t ) );
 		if ( data_size % ( params->fMtu - sizeof ( packet_header_t ) ) )
 			return ++npckt;
 		return npckt;
@@ -295,7 +295,7 @@ namespace Jack
 
 // Packet *******************************************************************************************************
 
-	EXPORT bool IsNextPacket ( packet_header_t* previous, packet_header_t* next, unsigned int subcycles )
+	EXPORT bool IsNextPacket ( packet_header_t* previous, packet_header_t* next, size_t subcycles )
 	{
 		//ignore first cycle
 		if ( previous->fCycle <= 1 )
